% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{apl}
\alias{apl}
\title{Full text of 'A pattern Language' by Christopher Alexander}
\description{
Full text of 'A pattern Language' by Christopher Alexander
}
\details{
each new pattern chapter beginning is marked with '>>>>>'.
subsections per pattern chapter are marked with '<<<<<'.
}
\author{
Christopher Alexander
}
\keyword{data}
