% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_functions.R
\name{apl_chapter_number}
\alias{apl_chapter_number}
\title{get graph vertices' chapter number}
\usage{
apl_chapter_number(g)
}
\arguments{
\item{graph}{igraph graph object}
}
\value{
integer vector with chapter numbers matching vertex order in g
}
\description{
get graph vertices' chapter number
}
