% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{chapter_headings}
\alias{chapter_headings}
\title{List of chapter headings in 'A pattern Language' by Christopher Alexander}
\description{
List of chapter headings in 'A pattern Language' by Christopher Alexander
}
\author{
Christopher Alexander
}
\keyword{data}
