% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bootstrap.R
\name{g_hypothesis_test_scalar_statistic}
\alias{g_hypothesis_test_scalar_statistic}
\title{Hypothesis test for graph statistics
apply a graph statistic to a graph, as well as randomised versions of it, and test the null-hypothesis that the graph statistic was produced by a random edge distribution.
Assumes normal distribution of the test statistic (returns normality test results along with p-values and results)}
\usage{
g_hypothesis_test_scalar_statistic(
  g,
  simulation.n = 100,
  scalar.graph.statistic,
  alpha = 0.05
)
}
\arguments{
\item{g}{igraph object}

\item{simulation.n}{how many simulations to run}

\item{scalar.graph.statistic}{a function that takes an igraph graph object as input and returns a scalar}

\item{alpha}{the confidence level}
}
\description{
Hypothesis test for graph statistics
apply a graph statistic to a graph, as well as randomised versions of it, and test the null-hypothesis that the graph statistic was produced by a random edge distribution.
Assumes normal distribution of the test statistic (returns normality test results along with p-values and results)
}
