% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_plot_functions.R
\name{g_plot}
\alias{g_plot}
\title{simple graph plotting}
\usage{
g_plot(g, interactive = T, ...)
}
\arguments{
\item{g}{the igraph object to plot}

\item{interactive}{logical; if TRUE, uses networkD3 for interactive animated plot}
}
\description{
simple graph plotting
}
