% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_plot_functions.R
\name{g_plot_degree_all}
\alias{g_plot_degree_all}
\title{plotting node degree (density distribution)}
\usage{
g_plot_degree_all(g)
}
\arguments{
\item{g}{igraph graph object}
}
\description{
plotting node degree (density distribution)
}
