% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_plot_functions.R
\name{g_plot_distance}
\alias{g_plot_distance}
\title{plot graph distance matrix}
\usage{
g_plot_distance(g, mode = "all")
}
\arguments{
\item{g}{igraph graph object}

\item{mode}{'in', 'out' or 'all' (for directed graphs)}
}
\description{
plot graph distance matrix
}
