% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bootstrap.R
\name{g_plot_hypothesis_test}
\alias{g_plot_hypothesis_test}
\title{histogram of graph hypothesis test}
\usage{
g_plot_hypothesis_test(
  g_hypothesis_test_scalar_statistic_result,
  alpha = 0.05,
  breaks = 30,
  col = "black",
  xlim = minmax(c(g_hypothesis_test_scalar_statistic_result$simulated,
    g_hypothesis_test_scalar_statistic_result$observed)),
  xlab = "statistic",
  ylab = "frequency",
  main = "degree variance (compared to randomised graphs)",
  sub = paste0("p <", ceiling(g_hypothesis_test_scalar_statistic_result$p.value *
    1000)/1000, "; assumption of normality",
    ifelse(g_hypothesis_test_scalar_statistic_result$normality.condition.met, " ",
    " not "), "met"),
  ...
)
}
\arguments{
\item{g_hypothesis_test_scalar_statistic_result}{result from g_hypothesis_test_scalar_statistic()}

\item{alpha}{what alpha to use to plot the 'plausible range'}

\item{...}{arguments passed to hist()}
}
\value{
the provided g_hypothesis_test_scalar_statistic_result as is. (function is only for plot side effect)
}
\description{
histogram of graph hypothesis test
}
