% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_plot_functions.R
\name{g_plot_interactive}
\alias{g_plot_interactive}
\title{simplified networkD3::forceNetwork() + with new defaults}
\usage{
g_plot_interactive(g, ..., groups = 1, weights = 1)
}
\arguments{
\item{g}{igraph graph object}
}
\description{
simplified networkD3::forceNetwork() + with new defaults
}
