% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_plot_functions.R
\name{g_plot_local}
\alias{g_plot_local}
\title{plotting the local part of a graph}
\usage{
g_plot_local(g, v, direction = "upstream", max.levels = 3, ...)
}
\arguments{
\item{g}{igraph graph object}
}
\description{
plotting the local part of a graph
}
