% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_plot_functions.R
\name{g_plot_vertex_stats_closeness_in_out}
\alias{g_plot_vertex_stats_closeness_in_out}
\title{plotting directed closeness centrality}
\usage{
g_plot_vertex_stats_closeness_in_out(g)
}
\arguments{
\item{g}{igraph graph object}
}
\description{
plotting directed closeness centrality
}
