% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_plot_functions.R
\name{g_plot_vertex_stats_degree_in_out}
\alias{g_plot_vertex_stats_degree_in_out}
\title{plotting node in/out degree (scatterplot)}
\usage{
g_plot_vertex_stats_degree_in_out(g, labels = F)
}
\arguments{
\item{g}{igraph graph object}

\item{labels}{if TRUE, points are labeled (not implemented)}
}
\description{
plotting node in/out degree (scatterplot)
}
