% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bootstrap.R
\name{g_randomise_edges_stat}
\alias{g_randomise_edges_stat}
\title{calculate graph statistics for randomised versions of a graph
unweighted, directed graphs only; graph is treated as simplified (no self-loops; no double edges)}
\usage{
g_randomise_edges_stat(g, simulation.n = 100, graph.statistic)
}
\description{
calculate graph statistics for randomised versions of a graph
unweighted, directed graphs only; graph is treated as simplified (no self-loops; no double edges)
}
