% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bootstrap.R
\name{g_randomised_statistic}
\alias{g_randomised_statistic}
\title{apply graph statistic to randomised copies of a graph}
\usage{
g_randomised_statistic(g, simulation.n, any.graph.statistic)
}
\arguments{
\item{g}{igraph object}

\item{simulation.n}{how many simulations to run}

\item{any.graph.statistic}{a function that takes a graph as input and returns any object}
}
\value{
a list with two items: 'simulation': a list, length equal 'simulation.n', with the resulting statistic for each randomised graph. 'original': the graph statistic for the original input graph
}
\description{
apply graph statistic to randomised copies of a graph
}
