% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bootstrap.R
\name{g_subgraph_giant_component}
\alias{g_subgraph_giant_component}
\title{Get largest connected component from graph}
\usage{
g_subgraph_giant_component(g, mode = "strong")
}
\arguments{
\item{g}{igraph garph}

\item{mode}{passed to igraph::components (whether edge direction is considered,defaults to 'strong' (alernative: 'weak'))}
}
\description{
Get largest connected component from graph
}
