% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_functions.R
\name{node_plant_node_ids}
\alias{node_plant_node_ids}
\title{Recursively identifying downstream nodes}
\usage{
node_plant_node_ids(
  g,
  v,
  direction = "downstream",
  max.levels = 2,
  level = 0,
  children_found = c()
)
}
\arguments{
\item{g}{igraph graph object}

\item{v}{the id of the central node to start from}

\item{direction}{should be 'downstream', 'upstream' or 'both'}

\item{max.levels}{how many steps away from source node to go}
}
\description{
Recursively identifying downstream nodes
}
