% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_functions.R
\name{node_river}
\alias{node_river}
\title{identifying downstream / upstream nodes}
\usage{
node_river(g, v, upstream = T, downstream = T, keep.bridges = T, ids_only = T)
}
\description{
identifying downstream / upstream nodes
}
