% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_functions.R
\name{nodeid}
\alias{nodeid}
\title{get the id of a a node in an igraph object by regex name search}
\usage{
nodeid(g, nodename)
}
\arguments{
\item{g}{igraph graph object}
}
\description{
get the id of a a node in an igraph object by regex name search
}
