% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl_functions.R
\name{page_rank_standardised}
\alias{page_rank_standardised}
\title{graph page rank with 'mode' argument}
\usage{
page_rank_standardised(g, mode = "all")
}
\arguments{
\item{g}{igraph graph object}

\item{mode}{mode must be "in", "out" or "all' (for directed pagerank)}
}
\description{
to match the other igraph vertex statistic functions
}
